week05/day2/DailyChallenge/phones/requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.0.2
phonenumbers==8.13.10
psycopg2==2.9.6
sqlparse==0.4.4
